requirements.txt
Django==3.2.12
djangorestframework==3.13.1
psycopg2==2.9.3
Pillow==9.1.0